match jirai_get_free_item_8006B268

+++ src/Bullet/jirai.h
@@ -38,7 +38,7 @@ typedef struct Actor_Jirai
   int field_128;
   int field_12C;
   int field_130;
-  int field_134_gcl_arg;
+  long field_134_gcl_arg;
   int field_138_gcl;
   int field_13C_idx;
   int field_140;
